[Fix-10827] Fix network error cause worker cannot send message to master (#10886) (#21)
* Fix network error cause worker cannot send message to master

(cherry picked from commit cade66a9b6b14450c6f778f7e05820b588856a41)

+++ dolphinscheduler-service/src/main/java/org/apache/dolphinscheduler/service/process/ProcessService.java
@@ -105,7 +105,7 @@ public interface ProcessService {
 
     void setSubProcessParam(ProcessInstance subProcessInstance);
 
-    TaskInstance submitTaskWithRetry(ProcessInstance processInstance, TaskInstance taskInstance, int commitRetryTimes, int commitInterval);
+    TaskInstance submitTaskWithRetry(ProcessInstance processInstance, TaskInstance taskInstance, int commitRetryTimes, long commitInterval);
 
     @Transactional
     TaskInstance submitTask(ProcessInstance processInstance, TaskInstance taskInstance);
